[#39] Separation of cli-client & service structure
