+++ cli-client/src/test/java/edu/kaist/algo/client/GcToolClientTest.java
@@ -13,8 +13,8 @@ import org.junit.runners.JUnit4;
  * Tests the functionality of parsing arguments for the client.
  */
 @RunWith(JUnit4.class)
-public class LogUploadParseTest {
-  private final Options options = LogUploadClient.makeOptions();
+public class GcToolClientTest {
+  private final Options options = GcToolClient.makeOptions();
 
   /**
    * Takes various combinations of options given, and tests for their validity.
@@ -22,11 +22,11 @@ public class LogUploadParseTest {
   @Test
   public void testParseOptions() {
     String[] args;
-    LogUploadClient.ParsedOptions parsedopt;
+    GcToolClient.ParsedOptions parsedopt;
 
     // working options
     args = new String[] {"-h", "localhost", "-p", "50051", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNotNull(parsedopt);
     assertEquals("localhost", parsedopt.getHost());
     assertEquals(50051, parsedopt.getPort());
@@ -34,7 +34,7 @@ public class LogUploadParseTest {
 
     // working options, given IP address as host
     args = new String[] {"-h", "148.255.255.1", "-p", "50051", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNotNull(parsedopt);
     assertEquals("148.255.255.1", parsedopt.getHost());
     assertEquals(50051, parsedopt.getPort());
@@ -42,55 +42,57 @@ public class LogUploadParseTest {
 
     // working options, implicit host = localhost
     args = new String[] {"-p", "50051", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNotNull(parsedopt);
     assertEquals("localhost", parsedopt.getHost());
     assertEquals(50051, parsedopt.getPort());
     assertEquals("logfile.log", parsedopt.getFilename());
 
-    // not existing filename option
+    // working options, implicit host = localhost, no file option
     args = new String[] {"-p", "50051"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
-    assertNull(parsedopt);
+    parsedopt = GcToolClient.parseOptions(options, args);
+    assertNotNull(parsedopt);
+    assertEquals("localhost", parsedopt.getHost());
+    assertEquals(50051, parsedopt.getPort());
 
     // not existing port number
     args = new String[] {"-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // invalid port number
     args = new String[] {"-p", "-1", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // invalid port number2
     args = new String[] {"-p", "99999", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // invalid port number3
     args = new String[] {"-p", "not a number", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // not existing option
     args = new String[] {"-u", "50051", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // not existing argument to an option
     args = new String[] {"-p", "-h"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // empty string argument
     args = new String[] {"-h", "", "-p", "50051", "-f", "logfile.log"};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
 
     // empty string argument 2
     args = new String[] {"-p", "50051", "-f", ""};
-    parsedopt = LogUploadClient.parseOptions(options, args);
+    parsedopt = GcToolClient.parseOptions(options, args);
     assertNull(parsedopt);
   }
 }
